+++ packages/orq-rc/src/funcs/deploymentsMetricsCreate.ts
@@ -29,6 +29,8 @@ import { Result } from "../types/fp.js";
  *
  * @remarks
  * Add metrics to a deployment
+ *
+ * @deprecated method: This will be removed in a future release, please migrate away from it as soon as possible.
  */
 export function deploymentsMetricsCreate(
   client: OrqCore,