+++ packages/orq-rc/RELEASES.md
@@ -7679,3 +7679,13 @@ Based on:
 - [typescript v4.6.0-rc.6] packages/orq-rc
 ### Releases
 - [NPM v4.6.0-rc.6] https://www.npmjs.com/package/@orq-ai/node/v/4.6.0-rc.6 - packages/orq-rc
+
+## 2026-03-12 22:14:58
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.753.0 (2.862.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.6.0-rc.8] packages/orq-rc
+### Releases
+- [NPM v4.6.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/4.6.0-rc.8 - packages/orq-rc